bees: separate BEES_MAX_EXTENT_TASK_COUNT from BEES_MAX_EXTENT_REF_COUNT
They are both roughly the right number, but the number of tasks is not the
same as the number of refs:

 * in extent scan mode, each extent Task has all of the extent's refs
 * in subvol scan mode, each Task handles inodes, not extents or refs

Making the extent task count smaller reduces parallel and out-of-order
task execution.  This helps prevent a wide gap from forming between the
current progress checkpoint (lowest extent in the queue) and the head
of the crawl (highest extent in the queue).  This makes restarts repeat
less work, which may help make forward progress if bees is run for very
short intervals between restarts, but it will make the scans take longer
due to more sequential processing.

Signed-off-by: Zygo Blaxell <bees@furryterror.org>